feat: autocomplete on search
* move csv atc files to json to use native imports rather than node:fs
* add MUI with DSFR integration for autocomplete
* separate DB types from connection
