+++ src/app/page.tsx
@@ -1,6 +1,5 @@
 import { fr } from "@codegouvfr/react-dsfr";
-import Input from "@codegouvfr/react-dsfr/Input";
-import Button from "@codegouvfr/react-dsfr/Button";
+import AutocompleteSearch from "@/components/AutocompleteSearch";
 
 export default async function Page() {
   return (
@@ -10,22 +9,7 @@ export default async function Page() {
           <h1 className={fr.cx("fr-h5")}>
             Quel médicament cherchez-vous&nbsp;?
           </h1>
-          <Input
-            label={"Quel médicament cherchez-vous&nbsp;?"}
-            hideLabel={true}
-            addon={
-              <Button
-                iconId={"fr-icon-search-line"}
-                title="Recherche"
-                type="submit"
-              />
-            }
-            nativeInputProps={{
-              placeholder: "Rechercher",
-              name: "s",
-              type: "search",
-            }}
-          />
+          <AutocompleteSearch />
         </form>
       </div>
     </div>
