+++ src/data/ATC-labels-1.json
@@ -1,0 +1,16 @@
+{
+  "A": "Système digestif",
+  "B": "Système sanguin",
+  "C": "Système cardiovasculaire",
+  "D": "Peau",
+  "G": "Système génital, urinaire et hormones sexuelles",
+  "H": "Hormones, thyroïde, corticoïdes",
+  "J": "Antibiotiques, anti-virus, vaccins",
+  "L": "Anti-cancéreux, système immunitaire",
+  "M": "Muscles et os",
+  "N": "Système nerveux",
+  "P": "Antiparasitaires, insecticides et répulsifs",
+  "R": "Système respiratoire",
+  "S": "Yeux et oreilles",
+  "V": "Divers"
+}
