+++ src/data/ATC-labels-2.json
@@ -1,0 +1,96 @@
+{
+  "A01": "Bouche et dents",
+  "A02": "Troubles de l'acidité",
+  "A03": "Troubles gastro-intestinaux",
+  "A04": "Nausée et vomissement",
+  "A05": "Foie et bile",
+  "A06": "Troubles du transit (constipation, diarrhée)",
+  "A07": "Troubles du transit (constipation, diarrhée)",
+  "A08": "Aide à la digestion",
+  "A09": "Aide à la digestion",
+  "A10": "Diabète",
+  "A11": "Vitamines, suppléments, stimulants",
+  "A12": "Vitamines, suppléments, stimulants",
+  "A13": "Vitamines, suppléments, stimulants",
+  "A14": "Vitamines, suppléments, stimulants",
+  "A15": "Vitamines, suppléments, stimulants",
+  "A16": "Aide à la digestion",
+  "B01": "Caillots sanguins",
+  "B02": "Hémorragie",
+  "B03": "Anémie",
+  "B05": "Perfusion sanguine",
+  "B06": "Perfusion sanguine",
+  "C01": "Médicaments pour le cœur",
+  "C02": "Régulation de la tension",
+  "C03": "Médicaments pour le cœur",
+  "C04": "Médicaments pour les vaisseaux sanguins",
+  "C05": "Médicaments pour les vaisseaux sanguins",
+  "C07": "Médicaments pour les vaisseaux sanguins",
+  "C08": "Médicaments pour les vaisseaux sanguins",
+  "C09": "Médicaments pour les vaisseaux sanguins",
+  "C10": "Statines (cholestérol)",
+  "D01": "Anti-champignons de peau",
+  "D02": "Crèmes protectrices et apaisantes",
+  "D03": "Crèmes protectrices et apaisantes",
+  "D04": "Crèmes protectrices et apaisantes",
+  "D05": "Psoriasis",
+  "D06": "Produits de chimiothérapie dermato",
+  "D07": "Cicatrisants",
+  "D08": "Antiseptiques et désinfectants",
+  "D09": "Antiseptiques et désinfectants",
+  "D10": "Médicaments contre l'acné",
+  "D11": "Autres produits dermato",
+  "G01": "Anti-infectieux et antiseptiques gynéco",
+  "G02": "Autres produits gynéco",
+  "G03": "Hormones sexuelles, contraceptifs, troubles de l'érection, prostate",
+  "G04": "Problèmes urinaires",
+  "H01": "Hormones",
+  "H02": "Corticoïdes",
+  "H03": "Médicaments de la thyroïde",
+  "H04": "Hormones",
+  "H05": "Médicaments de la thyroïde",
+  "J01": "Antibiotiques",
+  "J02": "Anti-champignons",
+  "J04": "Antibiotiques",
+  "J05": "Anti-virus",
+  "J06": "Immunité et anticorps",
+  "J07": "Vaccins",
+  "L01": "Anticancéreux",
+  "L02": "Endocrinologie",
+  "L03": "Stimulation du système immunitaire",
+  "L04": "Médicaments pour les maladies auto-immune, anti-rejet lors de greffes",
+  "M01": "Anti-inflammatoires et antirhumatismaux",
+  "M02": "Problèmes musculaires",
+  "M03": "Problèmes musculaires",
+  "M04": "Traitement de la goutte",
+  "M05": "Problèmes d'os",
+  "M09": "Problèmes d'os",
+  "N01": "Anesthésie",
+  "N02": "Anti-douleur",
+  "N03": "Anti-épileptiques",
+  "N04": "Maladie de Parkinson",
+  "N05": "Médicaments qui apaisent le psychisme",
+  "N06": "Antidépresseurs et stimulants",
+  "N07": "Troubles de l'addiction",
+  "P01": "Anti-parasites",
+  "P02": "Vermifuges",
+  "P03": "Répulsifs",
+  "R01": "Décongestionnants pour le nez",
+  "R02": "Médicaments pour la gorge",
+  "R03": "Voies respiratoires",
+  "R05": "Rhume et toux",
+  "R06": "Allergies",
+  "R07": "Autres médicaments du système respiratoire",
+  "S01": "Médicaments pour les yeux",
+  "S02": "Médicaments pour les oreilles",
+  "S03": "Préparations pour yeux et oreilles",
+  "V01": "Non classés",
+  "V03": "Non classés",
+  "V04": "Produits pour examens",
+  "V06": "Non classés",
+  "V07": "Produits pour examens",
+  "V08": "Produits de radio",
+  "V09": "Non classés",
+  "V10": "Non classés",
+  "V20": "Non classés"
+}
