+++ src/db/pdbmMySQL/types.ts
@@ -1,9 +1,6 @@
-import "server-cli-only";
+import { Selectable } from "kysely";
 
-import { Kysely, MysqlDialect, Selectable } from "kysely";
-import { createPool } from "mysql2";
-
-interface PdbmMySQL {
+export interface PdbmMySQL {
   Specialite: SpecialiteTable;
   Element: SpecElementTable;
   Composant: SpecComposantTable;
@@ -153,53 +150,3 @@ export enum PresentationComm {
   "Non communiquée" = 40,
   "Plus d'autorisation" = 45,
 }
-
-export const pdbmMySQL = new Kysely<PdbmMySQL>({
-  dialect: new MysqlDialect({
-    pool: process.env.PDBM_URL
-      ? createPool(process.env.PDBM_URL)
-      : createPool({
-          // .devcontainer config
-          database: "pdbm_bdd",
-          host: "db-mysql",
-          user: "root",
-          password: "mysql",
-          port: 3306,
-          connectionLimit: 10,
-        }),
-  }),
-});
-
-(async () => {
-  const StatutComm = await pdbmMySQL
-    .selectFrom("StatutComm")
-    .selectAll()
-    .execute();
-
-  const StatutAdm = await pdbmMySQL
-    .selectFrom("StatutAdm")
-    .selectAll()
-    .execute();
-
-  // Check that enums matches the database
-  Object.values(StatutComm).forEach(({ CommId, CommLibCourt, CommDomaine }) => {
-    if (
-      (CommDomaine === "S" ? SpecialiteComm : PresentationComm)[CommId] !==
-      CommLibCourt
-    ) {
-      throw new Error(
-        `Enum does not match database: ${CommId} ${CommLibCourt} ${CommDomaine}`,
-      );
-    }
-  });
-  Object.values(StatutAdm).forEach(({ StatId, StatLibCourt, StatDomaine }) => {
-    if (
-      (StatDomaine === "S" ? SpecialiteStat : PresentationStat)[StatId] !==
-      StatLibCourt
-    ) {
-      throw new Error(
-        `Enum does not match database: ${StatId} ${StatLibCourt} ${StatDomaine}`,
-      );
-    }
-  });
-})();
